requirements.txt
flask
pytesseract
requests
pillow
fpdf
python-docx
docx
